improve svelte library (#81)
* feat: Add initial files for Svelte library

This commit adds the initial files for a Svelte library, including configuration files (.eslintrc.cjs, .eslintignore, .gitignore, .npmrc, .prettierignore, and .prettierrc), package.json with devDependencies and peerDependencies, README.md file with basic instructions on how to create and develop a project using this library. It also includes new directories (src/lib/components/Layout and src/lib/components/Link) with their respective index.ts and svelte files. Finally, it adds an example page (+page.svelte) that uses the Layout component to display some text and an internal link.

* update root `package.json` with more+better keywords

* update eslint rules to allow svelte files

* replace old svelte-lib with new one

started with a fresh install of the lib based on the Svelte docs

* update root staging scripts to startup all libraries

* integrate svelte-lib with staging app

* better test name

* testing is all working!

* build fixed

* update all deps

* ignore linting on `/dist`

* resolve tsconfig issue with test files

* add basic a11y tests

* add library code coverage results

* 100% coverage

* svelte-lib // add tests to build script